Add new config enable-tcp-health-check
Currently the tcp health check is enabled based
on consul-notify, however not all the instances
of consul-client need the tcp health check.

This can be configured using juju integrations
by only ingegration consul-client with hypervisor
over consul-notify only if tcp health check is
required. However having relations in some instances
but not all of them might create confusion.

To keep it simple, a config option is
added in the charm to enable/disable tcp health
checks. For backward compatiblity, the config
option enable-tcp-health-check is set to true
by default.

Fixes: https://github.com/canonical/consul-client-operator/issues/21