Update healthcheck endpoint from consul-cluster relation data
conusl-k8s operator add a new key in relation data
gossip_healthcheck_endpoints which will be the loadbalancer
ip of the consul-server service with gossip port.
Use this endpoint to perform the health check.
It is expected that the endpoint will be in the same
network space as defined for this charm instance.
As a fallback, the current implementation of management
IPs will be used if healthcheck endpoint is not
present.

Rename the config enable-tcp-health-check to
enable-health-check as the consul config does
is adding health check with type script.